// front/src/backend_api/backend/models/EntryWithCategory.ts
export interface EntryWithCategory {
    /**
     * 
     * @type {string}
     * @memberof EntryWithCategory
     */
    categoryId: string;
    /**
     * 
     * @type {string}
     * @memberof EntryWithCategory
     */
    licenceNo: string;
    /**
     * 
     * @type {string}
     * @memberof EntryWithCategory
     */
    color: string | null;
    /**
     * 
     * @type {Date}
     * @memberof EntryWithCategory
     */
    registrationTime: Date | null;
    /**
     * 
     * @type {boolean}
     * @memberof EntryWithCategory
     */
    markedAsPresent: boolean | null;
    /**
     * 
     * @type {boolean}
     * @memberof EntryWithCategory
     */
    markedAsPaid: boolean;
    /**
     * 
     * @type {string}
     * @memberof EntryWithCategory
     */
    alternateName: string | null;
    /**
     * 
     * @type {number}
     * @memberof EntryWithCategory
     */
    minPoints: number;
    /**
     * 
     * @type {number}
     * @memberof EntryWithCategory
     */
    maxPoints: number;
    /**
     * 
     * @type {Date}
     * @memberof EntryWithCategory
     */
    startTime: Date;
    /**
     * 
     * @type {boolean}
     * @memberof EntryWithCategory
     */
    womenOnly: boolean;
    /**
     * 
     * @type {number}
     * @memberof EntryWithCategory
     */
    baseRegistrationFee: number;
    /**
     * 
     * @type {number}
     * @memberof EntryWithCategory
     */
    lateRegistrationFee: number;
    /**
     * 
     * @type {number}
     * @memberof EntryWithCategory
     */
    rewardFirst: number;
    /**
     * 
     * @type {number}
     * @memberof EntryWithCategory
     */
    rewardSecond: number;
    /**
     * 
     * @type {number}
     * @memberof EntryWithCategory
     */
    rewardSemi: number;
    /**
     * 
     * @type {number}
     * @memberof EntryWithCategory
     */
    rewardQuarter: number | null;
    /**
     * 
     * @type {number}
     * @memberof EntryWithCategory
     */
    maxPlayers: number;
    /**
     * 
     * @type {number}
     * @memberof EntryWithCategory
     */
    overbookingPercentage: number;
    /**
     * 
     * @type {number}
     * @memberof EntryWithCategory
     */
    rank: number;
}