// front/src/backend_api/backend/models/Player.ts
/* tslint:disable */
/* eslint-disable */
/**
 * USKB Tournament
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Gender } from './Gender';
import {
    GenderFromJSON,
    GenderFromJSONTyped,
    GenderToJSON,
    GenderToJSONTyped,
} from './Gender';

/**
 * 
 * @export
 * @interface Player
 */
export interface Player {
    /**
     * 
     * @type {string}
     * @memberof Player
     */
    licenceNo: string;
    /**
     * 
     * @type {string}
     * @memberof Player
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof Player
     */
    lastName: string;
    /**
     * 
     * @type {string}
     * @memberof Player
     */
    club: string;
    /**
     * 
     * @type {Gender}
     * @memberof Player
     */
    gender: Gender;
    /**
     * 
     * @type {number}
     * @memberof Player
     */
    nbPoints: number;
    /**
     * 
     * @type {number}
     * @memberof Player
     */
    bibNo?: number | null;
    /**
     * 
     * @type {string}
     * @memberof Player
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof Player
     */
    phone: string | null;
    /**
     * 
     * @type {number}
     * @memberof Player
     */
    totalActualPaid: number | null;
}



/**
 * Check if a given object implements the Player interface.
 */
export function instanceOfPlayer(value: object): value is Player {
    if (!('licenceNo' in value) || value['licenceNo'] === undefined) return false;
    if (!('firstName' in value) || value['firstName'] === undefined) return false;
    if (!('lastName' in value) || value['lastName'] === undefined) return false;
    if (!('club' in value) || value['club'] === undefined) return false;
    if (!('gender' in value) || value['gender'] === undefined) return false;
    if (!('nbPoints' in value) || value['nbPoints'] === undefined) return false;
    if (!('email' in value) || value['email'] === undefined) return false;
    if (!('phone' in value) || value['phone'] === undefined) return false;
    if (!('totalActualPaid' in value) || value['totalActualPaid'] === undefined) return false;
    return true;
}

export function PlayerFromJSON(json: any): Player {
    return PlayerFromJSONTyped(json, false);
}

export function PlayerFromJSONTyped(json: any, ignoreDiscriminator: boolean): Player {
    if (json == null) {
        return json;
    }
    return {
        
        'licenceNo': json['licenceNo'],
        'firstName': json['firstName'],
        'lastName': json['lastName'],
        'club': json['club'],
        'gender': GenderFromJSON(json['gender']),
        'nbPoints': json['nbPoints'],
        'bibNo': json['bibNo'] == null ? undefined : json['bibNo'],
        'email': json['email'],
        'phone': json['phone'],
        'totalActualPaid': json['totalActualPaid'],
    };
}

export function PlayerToJSON(json: any): Player {
    return PlayerToJSONTyped(json, false);
}

export function PlayerToJSONTyped(value?: Player | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'licenceNo': value['licenceNo'],
        'firstName': value['firstName'],
        'lastName': value['lastName'],
        'club': value['club'],
        'gender': GenderToJSON(value['gender']),
        'nbPoints': value['nbPoints'],
        'bibNo': value['bibNo'],
        'email': value['email'],
        'phone': value['phone'],
        'totalActualPaid': value['totalActualPaid'],
    };
}